requirements_test.txt
pytest==3.0.7
pytest-cov==2.5.1
python-dbusmock==0.16.7
mock==2.0.0
